chore: add parent activity component
