+++ bciers/apps/reporting/src/app/components/activities/ActivityForm.tsx
@@ -1,11 +1,6 @@
 "use client";
 import FormBase from "@bciers/components/form/FormBase";
-import FieldTemplate from "@bciers/components/form/fields/FieldTemplate";
-import NestedArrayFieldTemplate from "@bciers/components/form/fields/NestedArrayFieldTemplate";
-import SourceTypeBoxTemplate from "@bciers/components/form/fields/SourceTypeBoxTemplate";
-import CheckboxWidgetLeft from "@bciers/components/form/widgets/CheckboxWidgetLeft";
 import { customizeValidator } from "@rjsf/validator-ajv8";
-import InlineFieldTemplate from "@bciers/components/form/fields/InlineFieldTemplate";
 import { useEffect, useState } from "react";
 import { actionHandler } from "@bciers/actions";
 import { Alert, Button } from "@mui/material";
@@ -58,189 +53,14 @@ const data: TaskListElement[] = [
   { type: "Page", title: "Operation emission summary with a long title" },
 ];
 
-const uiSchema = {
-  "ui:FieldTemplate": FieldTemplate,
-  "ui:classNames": "form-heading-label",
-  gscWithProductionOfUsefulEnergy :{
-    "ui:FieldTemplate": FieldTemplate,
-    "ui:widget": CheckboxWidgetLeft,
-    "ui:options": {
-      label: false,
-    },
-  },
-  gscWithoutProductionOfUsefulEnergy :{
-    "ui:FieldTemplate": FieldTemplate,
-    "ui:widget": CheckboxWidgetLeft,
-    "ui:options": {
-      label: false,
-    },
-  },
-  sourceTypes: {
-    "ui:FieldTemplate": FieldTemplate,
-    "ui:options": {
-      label:false
-    },
-    gscWithProductionOfUsefulEnergy: {
-      "ui:FieldTemplate": SourceTypeBoxTemplate,
-      units: {
-        "ui:ArrayFieldTemplate": NestedArrayFieldTemplate,
-        "ui:FieldTemplate": FieldTemplate,
-        "ui:title": "Unit Data",
-        "ui:options": {
-          arrayAddLabel: 'Add Unit',
-          label: false,
-          title: 'Unit',
-          padding: 'p-2'
-        },
-        items:{
-          "ui:order": [
-            "gscUnitName",
-            "gscUnitType",
-            "description",
-            "fuels"
-          ],
-          gscUnitName: {
-            "ui:FieldTemplate": InlineFieldTemplate
-          },
-          gscUnitType: {
-            "ui:FieldTemplate": InlineFieldTemplate
-          },
-          description: {
-            "ui:FieldTemplate": InlineFieldTemplate
-          },
-          fuels: {
-            "ui:title": "Fuel Data",
-            "ui:ArrayFieldTemplate": NestedArrayFieldTemplate,
-            "ui:FieldTemplate": FieldTemplate,
-            "ui:options": {
-              arrayAddLabel: 'Add Fuel',
-              label: false,
-              title: 'Fuel'
-            },
-            items:{
-              "ui:order": [
-                "fuelName",
-                "fuelUnit",
-                "annualFuelAmount",
-                "emissions"
-              ],
-              fuelName: {
-                "ui:FieldTemplate": InlineFieldTemplate
-              },
-              fuelType: {
-                "ui:FieldTemplate": InlineFieldTemplate
-              },
-              annualFuelAmount: {
-                "ui:FieldTemplate": InlineFieldTemplate
-              },
-              emissions: {
-                "ui:ArrayFieldTemplate": NestedArrayFieldTemplate,
-                "ui:FieldTemplate": FieldTemplate,
-                "ui:options": {
-                  arrayAddLabel: 'Add Emission',
-                  title: 'Emission',
-                  label: false,
-                  verticalBorder: true
-                },
-                items: {
-                  "ui:order": [
-                    "gasType",
-                    "emission",
-                    "equivalentEmission",
-                    "methodology"
-                  ],
-                }
-              }
-            },
-          }
-        },
-      },
-    },
-    gscWithoutProductionOfUsefulEnergy: {
-      "ui:FieldTemplate": SourceTypeBoxTemplate,
-      units: {
-        "ui:ArrayFieldTemplate": NestedArrayFieldTemplate,
-        "ui:FieldTemplate": FieldTemplate,
-        "ui:title": "Unit Data",
-        "ui:options": {
-          arrayAddLabel: 'Add Unit',
-          title: 'Unit'
-        },
-        items:{
-          "ui:order": [
-            "gscUnitName",
-            "gscUnitType",
-            "description",
-            "fuels"
-          ],
-          "ui:options": {
-            title: 'Unit'
-          },
-          gscUnitName: {
-            "ui:FieldTemplate": InlineFieldTemplate
-          },
-          gscUnitType: {
-            "ui:FieldTemplate": InlineFieldTemplate
-          },
-          description: {
-            "ui:FieldTemplate": InlineFieldTemplate
-          },
-          fuels: {
-            "ui:title": "Fuel Data",
-            "ui:ArrayFieldTemplate": NestedArrayFieldTemplate,
-            "ui:FieldTemplate": FieldTemplate,
-            "ui:options": {
-              arrayAddLabel: 'Add Fuel'
-            },
-            items:{
-              "ui:order": [
-                "fuelName",
-                "fuelUnit",
-                "annualFuelAmount",
-                "emissions"
-              ],
-              fuelName: {
-                "ui:FieldTemplate": InlineFieldTemplate
-              },
-              fuelType: {
-                "ui:FieldTemplate": InlineFieldTemplate
-              },
-              annualFuelAmount: {
-                "ui:FieldTemplate": InlineFieldTemplate
-              },
-              emissions: {
-                "ui:ArrayFieldTemplate": NestedArrayFieldTemplate,
-                "ui:FieldTemplate": FieldTemplate,
-                "ui:options": {
-                  arrayAddLabel: 'Add Emission',
-                  title: 'Emission',
-                  label: false,
-                  verticalBorder: true
-                },
-                items: {
-                  "ui:order": [
-                    "gasType",
-                    "emission",
-                    "equivalentEmission",
-                    "methodology"
-                  ],
-                }
-              }
-            },
-          }
-        },
-      },
-    }
-  },
-};
-
 interface Props {
   activityData: {activityId: number, sourceTypeMap:{[key: number]: string}},
-  reportDate: string
+  reportDate: string,
+  uiSchema: any
 }
 
 // 🧩 Main component
-export default function ActivityForm({activityData, reportDate}:Readonly<Props>) {
+export default function ActivityForm({activityData, reportDate, uiSchema}:Readonly<Props>) {
   // 🐜 To display errors
   const [errorList, setErrorList] = useState([] as any[]);
   // 🌀 Loading state for the Submit button
@@ -266,7 +86,7 @@ export default function ActivityForm({activityData, reportDate}:Readonly<Props>)
       // fetch data from server
       const schemaData = await actionHandler(`reporting/build-form-schema?activity=${activityId}&report_date=${reportDate}${selectedSourceTypes}`, "GET", "");
       setJsonSchema(JSON.parse(schemaData).schema);
-      setFormState({...formState, sourceTypes: {gscWithProductionOfUsefulEnergy: {units: [{fuels: [{emissions:[{}]}]}]}}})
+      setFormState({...formState, sourceTypes: {gscWithProductionOfUsefulEnergy: {units: [{fuels: [{emissions:[{}]}]}]}, gscWithoutProductionOfUsefulEnergy: {units: [{fuels: [{emissions:[{}]}]}]}}})
     };
     let selectedSourceTypes = '';
     for (const [key, value] of Object.entries(sourceTypeMap)) {
@@ -276,10 +96,8 @@ export default function ActivityForm({activityData, reportDate}:Readonly<Props>)
   }, dependencyArray)
 
   const customFormats = {
-    phone: /\(?\d{3}\)?[\s-]?\d{3}[\s-]?\d{4}$/,
-    "postal-code":
-      /^[ABCEGHJ-NPRSTVXY]\d[ABCEGHJ-NPRSTV-Z][ -]?\d[ABCEGHJ-NPRSTV-Z]\d$/i,
-    bc_corporate_registry_number: "^[A-Za-z]{1,3}\\d{7}$",
+    // Add any needed custom formats here like the example below
+    // phone: /\(?\d{3}\)?[\s-]?\d{3}[\s-]?\d{4}$/,
   };
 
   const validator = customizeValidator({ customFormats });
