+++ bciers/apps/reporting/src/app/demo/activities/general_stationary_combustion/page.tsx
@@ -1,4 +1,4 @@
-import ActivityForm from "../../../components/activities/ActivityForm";
+import GeneralStationaryCombustion from "../../../components/activities/generalStationaryCombustion";
 import { Suspense } from "react";
 import { actionHandler } from "@bciers/actions";
 
@@ -6,9 +6,10 @@ export default async function Page() {
   const reportDate = '2024-04-01' // This should be passed in once we have a path to this page from starting a report
   const activityData = await actionHandler(`reporting/get-initial-activity-data?activity_name=General stationary combustion&report_date=${reportDate}`, "GET", "");
   const activityDataObject = JSON.parse(activityData);
+
   return (
-  <Suspense fallback="Loading Schema">
-    <ActivityForm activityData={activityDataObject} reportDate={reportDate}/>
-  </Suspense>
+    <Suspense fallback="Loading Schema">
+      <GeneralStationaryCombustion activityData={activityDataObject} reportDate={reportDate}/>
+    </Suspense>
   )
 }
